backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
bcrypt
pyjwt
python-multipart
pydantic[email]
httpx
python-dotenv
